Implemented correction for 12pm solar noon
Program checks at the end of every day whether the solar noon matches 12pm, and if this is not the case, it adjusts the time by jumping forward or back to account for this. This does not take into account that solar noon is at 1pm during daylight savings time. It also relies on the fact that the sun rises and sets at some point during the day.